+++ web-console/src/components/Event.tsx
@@ -3,8 +3,8 @@ import type { FC } from "hono/jsx";
 import type { DateTime } from "../factory";
 import { Button } from "./Button";
 import { Description, DescriptionList } from "./DescriptionList";
-import { ScanSearch, Sunrise } from "./Icon";
-import { SharedModalContent, type SharedModalData } from "./Modal";
+import { ScanSearch, Spinner, SunMedium, Sunrise } from "./Icon";
+import { SharedModalContent } from "./Modal";
 import { StatusIndicator } from "./StatusIndicator";
 import {
   Table,
@@ -20,25 +20,13 @@ import type { EventWithDispatches } from "./types";
 export const Event: FC<{
   event: EventWithDispatches;
   formatDate: (d: DateTime) => string;
-  sharedModal: SharedModalData;
-}> = ({ event, formatDate, sharedModal }) => {
-  let eventStatus: "ongoing" | "complete" | "ignored" | "failed" = "ongoing";
-  const statuses = event.dispatches
-    .map((d) => d.status)
-    .filter((s) => s !== "ignored");
-  if (statuses.length === 0) {
-    eventStatus = "ignored";
-  } else if (statuses.some((s) => s === "ongoing")) {
-    eventStatus = "ongoing";
-  } else if (statuses.every((s) => s === "complete")) {
-    eventStatus = "complete";
-  } else if (
-    statuses.some(
-      (s) => s === "failed" || s === "misconfigured" || s === "lost",
-    )
-  ) {
-    eventStatus = "failed";
-  }
+  refreshIntervalSeconds: number;
+  eventTitle?: (e: EventWithDispatches) => string;
+}> = ({ event, formatDate, refreshIntervalSeconds, eventTitle }) => {
+  const title = eventTitle ? eventTitle(event) : event.id;
+  const payload = JSON.stringify(event.payload, null, 4);
+  const rows = payload.split("\n").length;
+
   return (
     <div
       id={`event-${event.id}`}
@@ -46,16 +34,33 @@ export const Event: FC<{
       hx-swap-oob={`#event-${event.id}`}
     >
       <div class="flex justify-between font-semibold leading-7">
-        <div class="flex place-items-center gap-2">
-          <StatusIndicator status={eventStatus} />
-          <p class="text-gray-900">{formatDate(event.createdAt)}</p>
+        <div class="flex place-items-center gap-1">
+          <SunMedium title="" />
+          <p class="text-gray-900">{title}</p>
         </div>
-        <p class="text-gray-500">{event.id}</p>
+        <p class="text-gray-500">{formatDate(event.createdAt)}</p>
       </div>
       <div class="relative my-4 text-gray-500 flex flex-col]">
-        <Textarea readonly>{JSON.stringify(event.payload, null, 4)}</Textarea>
+        <Textarea rows={rows} readonly>
+          {payload}
+        </Textarea>
         <button
-          class="absolute right-0 px-1 mx-1 my-1 rounded-md hover:bg-gray-200 active:outline active:outline-1 active:bg-gray-400 active:text-white"
+          class="
+            absolute
+            right-0
+            px-2
+            py-1
+            mx-1
+            my-1
+            rounded-md
+            hover:bg-gray-200
+            active:outline
+            active:outline-1
+            active:bg-gray-400
+            active:text-white
+            text-sm
+            select-none
+          "
           type="button"
           data-payload={JSON.stringify(event.payload)}
           _="on click
@@ -80,85 +85,113 @@ export const Event: FC<{
             <TableHead>Details</TableHead>
           </TableRow>
         </TableHeader>
-        <TableBody>
-          {event.dispatches.length > 0 ? (
-            event.dispatches.map((dispatch) => (
-              <Dispatch
-                key={dispatch.id}
-                dispatch={dispatch}
-                formatDate={formatDate}
-                sharedModal={sharedModal}
-              />
-            ))
-          ) : (
-            <TableRow>
-              <TableCell class="text-center pt-6 pb-2" colspan={5}>
-                no dispatches.
-              </TableCell>
-            </TableRow>
-          )}
-        </TableBody>
+        <EventDispatches
+          eventId={event.id}
+          dispatches={event.dispatches}
+          formatDate={formatDate}
+          refreshIntervalSeconds={refreshIntervalSeconds}
+        />
       </Table>
     </div>
+  );
+};
+
+export const EventDispatches = ({
+  eventId,
+  dispatches,
+  formatDate,
+  refreshIntervalSeconds,
+}: {
+  eventId: string;
+  dispatches: EventWithDispatches["dispatches"];
+  formatDate: (d: DateTime) => string;
+  refreshIntervalSeconds: number;
+}) => {
+  const id = `event-dispatches-${eventId}`;
+  const inProgress = dispatches.some((d) => d.status === "ongoing");
+  const updateAttributes = inProgress
+    ? {
+        "hx-get": `/components/event/${encodeURIComponent(eventId)}/dispatches`,
+        "hx-target": "this",
+        "hx-swap": "outerHTML",
+        "hx-trigger": `every ${refreshIntervalSeconds}s[document.visibilityState === 'visible']`,
+      }
+    : {};
+
+  return (
+    <TableBody id={id} {...updateAttributes}>
+      {dispatches.length > 0 ? (
+        dispatches.map((dispatch) => (
+          <Dispatch
+            key={dispatch.id}
+            dispatch={dispatch}
+            formatDate={formatDate}
+          />
+        ))
+      ) : (
+        <TableRow>
+          <TableCell class="text-center pt-6 pb-2" colspan={5}>
+            no dispatches.
+          </TableCell>
+        </TableRow>
+      )}
+    </TableBody>
   );
 };
 
 export const Dispatch: FC<{
   dispatch: EventWithDispatches["dispatches"][number];
   formatDate: (d: DateTime) => string;
-  sharedModal: SharedModalData;
-}> = ({ dispatch, formatDate, sharedModal }) => (
-  <TableRow>
-    <TableCell>
-      <code>{dispatch.destination}</code>
-    </TableCell>
-    <TableCell>
-      <div class="flex gap-1 items-center">
-        <StatusIndicator status={dispatch.status} />
-        {dispatch.status}
-      </div>
-    </TableCell>
-    <TableCell>
-      {dispatch.executionLog.length} / {dispatch.maxRetries + 1}
-    </TableCell>
-    <TableCell>
-      {formatDate(
-        dispatch.executionLog.length > 0
-          ? dispatch.executionLog[dispatch.executionLog.length - 1].executedAt
-          : dispatch.createdAt,
-      )}
-    </TableCell>
-    <TableCell>
-      <SharedModalContent
-        sharedModal={sharedModal}
-        contentFrameId={`dispatch-${dispatch.id}`}
-        trigger={(openModal) => (
+}> = ({ dispatch, formatDate }) => {
+  const indicatorId = `loading-${dispatch.id}`;
+  return (
+    <TableRow>
+      <TableCell>
+        <code>{dispatch.destination}</code>
+      </TableCell>
+      <TableCell>
+        <div class="flex gap-1 items-center">
+          <StatusIndicator status={dispatch.status} />
+          {dispatch.status}
+        </div>
+      </TableCell>
+      <TableCell>
+        {dispatch.executionLog.length} / {dispatch.maxRetries + 1}
+      </TableCell>
+      <TableCell>
+        {formatDate(
+          dispatch.executionLog.length > 0
+            ? dispatch.executionLog[dispatch.executionLog.length - 1].executedAt
+            : dispatch.createdAt,
+        )}
+      </TableCell>
+      <TableCell>
+        <SharedModalContent
+          contentFrameId={`dispatch-${dispatch.id}`}
+          trigger={(openModal) => (
+            <div class="w-fit cursor-pointer hover:text-gray-500" _={openModal}>
+              <ScanSearch title="Show detail" />
+            </div>
+          )}
+        >
           <div
-            class="w-fit cursor-pointer hover:text-gray-500"
-            hx-swap-oob={`.dispatch-${dispatch.id}`}
-            _={openModal}
+            hx-get={`/components/dispatch-detail-modal/${encodeURIComponent(dispatch.id)}`}
+            hx-trigger="intersect"
+            hx-swap="innerHTML"
+            hx-indicator={`#${indicatorId}`}
           >
-            <ScanSearch title="Show detail" />
+            <Spinner />
           </div>
-        )}
-      >
-        {(closeModal) => (
-          <DispatchDetails
-            dispatch={dispatch}
-            closeModal={closeModal}
-            formatDate={formatDate}
-          />
-        )}
-      </SharedModalContent>
-    </TableCell>
-  </TableRow>
-);
+        </SharedModalContent>
+      </TableCell>
+    </TableRow>
+  );
+};
 
-const DispatchDetails: FC<{
+export const DispatchDetails: FC<{
   dispatch: EventWithDispatches["dispatches"][number];
-  closeModal: string;
   formatDate: (d: DateTime) => string;
-}> = ({ dispatch, closeModal, formatDate }) => (
+}> = ({ dispatch, formatDate }) => (
   <div class={`dispatch-${dispatch.id}`}>
     <h2 class="text-2xl font-semibold">
       <span class="flex gap-2 items-center">
@@ -201,7 +234,7 @@ const DispatchDetails: FC<{
               <TableRow>
                 <TableHead>No.</TableHead>
                 <TableHead>Result</TableHead>
-                <TableHead>ExecutedAt</TableHead>
+                <TableHead>Executed at</TableHead>
               </TableRow>
             </TableHeader>
             <TableBody>
@@ -226,7 +259,15 @@ const DispatchDetails: FC<{
       >
         Retry as new dispatch
       </Button>
-      <Button type="button" _={closeModal} secondary>
+      <Button
+        type="button"
+        _="
+          on click
+            set dialog to the closest <dialog/>
+            call dialog.close()
+        "
+        secondary
+      >
         Close
       </Button>
     </div>